+++ SavingsGoalManager/src/main/java/com/brenner/budgetmanager/savingsgoals/SavingsGoal.java
@@ -3,6 +3,7 @@ package com.brenner.budgetmanager.savingsgoals;
 import org.springframework.format.annotation.DateTimeFormat;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -26,7 +27,7 @@ public class SavingsGoal {
     @DateTimeFormat(pattern = "yyyy-MM-dd")
     private Date savingsEndDate;
     
-    private Float targetAmount;
+    private BigDecimal targetAmount;
     
     private Integer monthsTillPayment;
     
@@ -34,15 +35,15 @@ public class SavingsGoal {
     
     private Integer daysTillPayment;
     
-    private Float savingsPerMonth;
+    private BigDecimal savingsPerMonth;
     
-    private Float savingsPerWeek;
+    private BigDecimal savingsPerWeek;
     
-    private Float savingsPerDay;
+    private BigDecimal savingsPerDay;
     
-    private Float initialBalance;
+    private BigDecimal initialBalance;
     
-    private Float currentBalance;
+    private BigDecimal currentBalance;
     
     private boolean isDefault;
     
@@ -50,8 +51,8 @@ public class SavingsGoal {
     
     public SavingsGoal() {}
     
-    public SavingsGoal(Integer savingsGoalId, String goalName, Date savingsStartDate, Date savingsEndDate, 
-    		Float targetAmount, Float initialBalance, Float currentBalance, boolean isDefault) {
+    public SavingsGoal(Integer savingsGoalId, String goalName, Date savingsStartDate, Date savingsEndDate,
+                       BigDecimal targetAmount, BigDecimal initialBalance, BigDecimal currentBalance, boolean isDefault) {
     	this.savingsGoalId = savingsGoalId;
     	this.goalName = goalName;
     	this.savingsStartDate = savingsStartDate;
@@ -94,11 +95,11 @@ public class SavingsGoal {
         this.savingsEndDate = savingsEndDate;
     }
 
-    public Float getTargetAmount() {
+    public BigDecimal getTargetAmount() {
         return targetAmount;
     }
 
-    public void setTargetAmount(Float targetAmount) {
+    public void setTargetAmount(BigDecimal targetAmount) {
         this.targetAmount = targetAmount;
     }
 
@@ -126,35 +127,35 @@ public class SavingsGoal {
         this.daysTillPayment = daysTillPayment;
     }
 
-    public Float getSavingsPerMonth() {
+    public BigDecimal getSavingsPerMonth() {
         return savingsPerMonth;
     }
 
-    public void setSavingsPerMonth(Float savingsPerMonth) {
+    public void setSavingsPerMonth(BigDecimal savingsPerMonth) {
         this.savingsPerMonth = savingsPerMonth;
     }
 
-    public Float getSavingsPerWeek() {
+    public BigDecimal getSavingsPerWeek() {
         return savingsPerWeek;
     }
 
-    public void setSavingsPerWeek(Float savingsPerWeek) {
+    public void setSavingsPerWeek(BigDecimal savingsPerWeek) {
         this.savingsPerWeek = savingsPerWeek;
     }
 
-    public Float getSavingsPerDay() {
+    public BigDecimal getSavingsPerDay() {
         return savingsPerDay;
     }
 
-    public void setSavingsPerDay(Float savingsPerDay) {
+    public void setSavingsPerDay(BigDecimal savingsPerDay) {
         this.savingsPerDay = savingsPerDay;
     }
 
-    public Float getCurrentBalance() {
+    public BigDecimal getCurrentBalance() {
 		return currentBalance;
 	}
 
-	public void setCurrentBalance(Float currentBalance) {
+	public void setCurrentBalance(BigDecimal currentBalance) {
 		this.currentBalance = currentBalance;
 	}
 
@@ -174,11 +175,11 @@ public class SavingsGoal {
         this.notes = notes;
     }
 
-    public Float getInitialBalance() {
+    public BigDecimal getInitialBalance() {
 		return initialBalance;
 	}
 
-	public void setInitialBalance(Float initialBalance) {
+	public void setInitialBalance(BigDecimal initialBalance) {
 		this.initialBalance = initialBalance;
 	}
